change json loading
