+++ beacon-chain/casper/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "//shared/bitutil:go_default_library",
         "//shared/mathutil:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
-        "@com_github_golang_protobuf//jsonpb:go_default_library_gen",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
