+++ beacon-chain/casper/validator.go
@@ -3,9 +3,7 @@ package casper
 import (
 	"bytes"
 	"fmt"
-	"os"
 
-	"github.com/golang/protobuf/jsonpb"
 	"github.com/prysmaticlabs/prysm/beacon-chain/params"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
 	"github.com/prysmaticlabs/prysm/shared/bitutil"
@@ -27,25 +25,6 @@ func InitialValidators() []*pb.ValidatorRecord {
 		validators = append(validators, validator)
 	}
 	return validators
-}
-
-// InitialValidatorsFromJSON retrieves the validator set that is stored in
-// genesis.json.
-func InitialValidatorsFromJSON(genesisJSONPath string) ([]*pb.ValidatorRecord, error) {
-	// #nosec G304
-	// genesisJSONPath is a user input for the path of genesis.json.
-	// Ex: /path/to/my/genesis.json.
-	f, err := os.Open(genesisJSONPath)
-	if err != nil {
-		return nil, err
-	}
-
-	cState := &pb.CrystallizedState{}
-	if err := jsonpb.Unmarshal(f, cState); err != nil {
-		return nil, fmt.Errorf("error converting JSON to proto: %v", err)
-	}
-
-	return cState.Validators, nil
 }
 
 // ActiveValidatorIndices filters out active validators based on start and end dynasty
