+++ beacon-chain/utils/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "clock.go",
         "flags.go",
+        "genesis_json.go",
         "shuffle.go",
         "slot_ticker.go",
         "vote_cache.go",
@@ -13,7 +14,9 @@ go_library(
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
         "//beacon-chain/params:go_default_library",
+        "//proto/beacon/p2p/v1:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
+        "@com_github_golang_protobuf//jsonpb:go_default_library_gen",
         "@com_github_urfave_cli//:go_default_library",
         "@org_golang_x_crypto//blake2b:go_default_library",
     ],
